requirements.txt
langchain ==0.0.352
openai ==0.27.8
python-dotenv ==1.0.0
tiktoken
chromadb
pyboxen
pandas
pyfiglet